requirements.txt
streamlit
langchain
langchain-openai
langchain-core
langgraph
python-docx
PyPDF2
streamlit-authenticator
pyyaml
paypalrestsdk